Comments added: functionality for 0-3 rotors fixed To do: match the descriptive error messages to the ones on LabTS

+++ Plugboard.cpp
@@ -31,10 +31,10 @@ int Plugboard::load_plugboard(char filename[20]) {
 
     vector<int> inputInts;
 
-    for (int i = 0; i < inputWords.size(); i++) {
+    for (size_t i = 0; i < inputWords.size(); i++) {
         try {inputInts.push_back(stoi(inputWords[i]));} // Convert vector of strings to vector of ints
         catch (invalid_argument &e) {return NON_NUMERIC_CHARACTER;} // Catch in case vector cannot be converted to int
-        for (int j = 0; j < inputWords[i].length(); j++) {
+        for (size_t j = 0; j < inputWords[i].length(); j++) {
             if (!isdigit(inputWords[i][j])) {
                 return NON_NUMERIC_CHARACTER;
             }
@@ -42,7 +42,7 @@ int Plugboard::load_plugboard(char filename[20]) {
         if (inputInts[i] < 0 || inputInts[i] >= 26)
             return INVALID_INDEX;
         if (i != 0) {
-            for (int j = 0; j < i-1; j++) {
+            for (size_t j = 0; j < i-1; j++) {
                 if (inputInts[i] == inputInts[j]) {
                     return IMPOSSIBLE_PLUGBOARD_CONFIGURATION;
                 }
@@ -63,14 +63,14 @@ void Plugboard::insert_to_vector(vector<int>& plugboard_vector, string contents)
         inputWords.push_back(word);
     }
 
-    for (int i = 0; i < inputWords.size(); i++) {
+    for (size_t i = 0; i < inputWords.size(); i++) {
         plugboard_vector.push_back(stoi(inputWords[i]));
     }
 }
 
 void Plugboard::print_ints(vector<int> plugboard_vector) {
     cout << "The plugboard integers loaded are:\n";
-    for (int i = 0; i < plugboard_vector.size(); i++) {
+    for (size_t i = 0; i < plugboard_vector.size(); i++) {
         cout << plugboard_vector[i] << " ";
     }
     cout << endl;
